reduce info type size

+++ lib/SPVM/Builder/include/spvm_runtime_info_type.h
@@ -6,8 +6,8 @@
 // Field information
 struct SPVM_runtime_info_type {
   int32_t basic_type_id;
-  int32_t dimension;
-  int32_t flag;
+  uint16_t dimension;
+  uint16_t flag;
 };
 
 SPVM_RUNTIME_INFO_TYPE* SPVM_RUNTIME_INFO_TYPE_new();
